% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{anova_tbl}
\alias{anova_tbl}
\title{One-way ANOVA analysis}
\usage{
anova_tbl(
  data,
  outcome_var,
  group_var,
  var_equal = TRUE,
  outcome_label = NULL,
  group_label = "Groups",
  table_caption = "Comparison of means between groups",
  posthoc_method = "bonferroni",
  abbreviation = NULL
)
}
\arguments{
\item{data}{Data frame}

\item{outcome_var}{Continuous outcome variable}

\item{group_var}{Grouping variable with more than 2 groups}

\item{var_equal}{TRUE(ANOVA) FALSE(Welch ANOVA)}

\item{outcome_label}{New label for outcome variable}

\item{group_label}{Header for the column listing the groups}

\item{table_caption}{Caption for the table in string}

\item{posthoc_method}{P-value adjustment for post-hoc test}

\item{abbreviation}{Full name of abbreviated variables}
}
\value{
A One-Way ANOVA output table
}
\description{
A One-Way ANOVA output table with post-hoc adjustment
}
\examples{
anova_tbl(
  data = mjms_data,
  outcome_var = "Cholesterol",
  group_var = "Treatment_Group",
  outcome_label = "Serum Cholesterol"
)

}
