% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{chisq_tbl}
\alias{chisq_tbl}
\title{Chi-square test}
\usage{
chisq_tbl(
  data,
  exposure_var,
  outcome_var,
  table_caption = "Association between exposure and outcome"
)
}
\arguments{
\item{data}{Data frame}

\item{exposure_var}{Categorical exposure variable}

\item{outcome_var}{Categorical outcome variable}

\item{table_caption}{Caption for the table in string}
}
\value{
A Chi-square output table
}
\description{
A Chi-square output table
}
\examples{
chisq_tbl(
  data = mjms_data,
  exposure_var = "Sex",
  outcome_var = "Smoker",
  table_caption = "Association of Gender vs. Smoking Status."
)

}
