% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{logistic_tbl}
\alias{logistic_tbl}
\title{Logistic regression test}
\usage{
logistic_tbl(
  data,
  outcome_var,
  predictor_vars,
  ref_levels = NULL,
  var_labels = NULL,
  cat_vars = NULL,
  table_caption = "Associated factors",
  abbreviation = NULL
)
}
\arguments{
\item{data}{Data frame}

\item{outcome_var}{Binary output variable as a string}

\item{predictor_vars}{Predictor variables as a character vector. Interaction
terms can be included using standard R formula syntax, e.g. "Age * Sex"
or "Age:Sex"}

\item{ref_levels}{List name of reference levels (optional)}

\item{var_labels}{List name of variable labels}

\item{cat_vars}{Categorical variables that wish to see the reference variable}

\item{table_caption}{Caption for the table in string}

\item{abbreviation}{Full name of abbreviated variables}
}
\value{
A logistic regression output table
}
\description{
A logistic regression output table
}
\examples{
logistic_tbl(
data = mjms_data,
outcome_var = "Heart_Disease",
predictor_vars = c("Age * Sex", "Smoker"), # "Age * Sex" adds an interaction term
ref_levels = list(Smoker = "No", Sex = "Female"),
cat_vars = c("Sex", "Smoker"),
table_caption = "Risk Factors for Heart Disease"
)

}
