% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{mcnemar_tbl}
\alias{mcnemar_tbl}
\title{McNemar test}
\usage{
mcnemar_tbl(
  data,
  pre_var,
  post_var,
  table_caption = "Association between pre and post",
  header_pre = "Pre",
  header_post = "Post"
)
}
\arguments{
\item{data}{Data frame}

\item{pre_var}{Pre variable}

\item{post_var}{Post variable}

\item{table_caption}{Caption for the table in string}

\item{header_pre}{Header for row variable}

\item{header_post}{Header for column variable}
}
\value{
A Mc Nemar output table
}
\description{
A MC Nemar output table
}
\examples{
mcnemar_tbl(
data = mjms_data,
pre_var = "Symptom_Pre",
post_var = "Symptom_Post",
table_caption = "Symptom status pre- and post-treatment.",
header_pre = "Pre Symptom",
header_post = "Post Symptom"
)

}
