% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MJMSTable-package.R
\docType{package}
\name{MJMSTable-package}
\alias{MJMSTable}
\alias{MJMSTable-package}
\title{MJMSTable: Produce MJMS Statistical Tables using R}
\description{
R package for generating publication-ready tables according to The Malaysian Journal of Medical Sciences (MJMS, \url{http://www.mjms.usm.my/index.html}) statistical requirements \url{http://www.mjms.usm.my/MJMS23052016/01MJMS23052016_ED.pdf}.
}
\author{
\strong{Maintainer}: Wan Nor Arifin \email{wnarifin@gmail.com} (\href{https://orcid.org/0000-0001-7786-4251}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Wan Nor Arifin \email{wnarifin@gmail.com} (\href{https://orcid.org/0000-0001-7786-4251}{ORCID}) [copyright holder]
  \item Ainin Sofiya Mastura Shamsudin \email{aininsofiya828@gmail.com}
}

}
\keyword{internal}
