% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{paired_ttest_tbl}
\alias{paired_ttest_tbl}
\title{Paired t-test}
\usage{
paired_ttest_tbl(
  data,
  id_var,
  pre_var,
  post_var,
  variable_label = NULL,
  table_caption = "Comparison between paired group",
  abbreviation = NULL
)
}
\arguments{
\item{data}{Data frame}

\item{id_var}{ID variable or unique identifier}

\item{pre_var}{Pre variable continuous}

\item{post_var}{Post variable continuous}

\item{variable_label}{New label for outcome variable}

\item{table_caption}{Caption for the table in string}

\item{abbreviation}{Full name of abbreviated variables}
}
\value{
A Paired t-test output table
}
\description{
A Paired t-test output table
}
\examples{
paired_ttest_tbl(
  data = mjms_data,
  id_var = "PatientID",
  pre_var = "SBP_Left_Arm",
  post_var = "SBP_Right_Arm",
  variable_label = "Systolic Blood Pressure (mmHg)",
  table_caption = "Comparison of SBP between Left and Right Arms"
)

}
