% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{ttest_tbl}
\alias{ttest_tbl}
\title{Independent t-test}
\usage{
ttest_tbl(
  data,
  outcome_var,
  group_var,
  equal_var = TRUE,
  table_caption = "Comparison between groups",
  abbreviation = NULL
)
}
\arguments{
\item{data}{Data frame}

\item{outcome_var}{Continuous output variable}

\item{group_var}{Grouping variable consisting of two levels}

\item{equal_var}{TRUE(t-test) FALSE(Welch t-test)}

\item{table_caption}{Caption for the table in string}

\item{abbreviation}{Full name of abbreviated variables}
}
\value{
An independent t-test output table
}
\description{
An Independent t-test output table
}
\examples{
ttest_tbl(
  data = mjms_data,
  outcome_var = "BMI",
  group_var = "Sex",
  equal_var = TRUE,
  table_caption = "Comparison of BMI by Biological Sex",
  abbreviation = "BMI = body mass index"
)

}
