% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{descriptive_tbl}
\alias{descriptive_tbl}
\title{Descriptive statistic test}
\usage{
descriptive_tbl(
  data,
  group_var,
  included_var,
  continuous_vars = NULL,
  table_caption = "Patient Demographics",
  abbreviation = NULL
)
}
\arguments{
\item{data}{Data frame}

\item{group_var}{Grouping variable}

\item{included_var}{Variable to be included from the data frame}

\item{continuous_vars}{Variables to be presented as continuous (mean and SD)}

\item{table_caption}{Caption for the table in string}

\item{abbreviation}{Full name of abbreviated variables}
}
\value{
A descriptive table
}
\description{
A descriptive table describing demographic or simple variable
}
\examples{
descriptive_tbl(
  data = mjms_data,
  group_var = "Treatment_Group",
  included_var = c("Age", "BMI", "Sex", "Smoker"),
  table_caption = "Baseline Patient Characteristics",
  abbreviation = "BMI = Body Mass Index, SBP = Systolic Blood Pressure"
)
}
