% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{diagnostic_tbl}
\alias{diagnostic_tbl}
\title{Diagnostic test}
\usage{
diagnostic_tbl(
  data,
  status_var,
  marker_map,
  show_plot = FALSE,
  plot_marker = NULL,
  table_caption = "Diagnostic performance",
  abbreviation = NULL
)
}
\arguments{
\item{data}{Data frame}

\item{status_var}{Target variable (Gold standard)}

\item{marker_map}{Define index tests/markers, friendly labels and cutoff points for every markers. See example below for settings.}

\item{show_plot}{TRUE(plot) FALSE(no plot)}

\item{plot_marker}{Specific test visualize in the ROC plot}

\item{table_caption}{Caption for the table in string}

\item{abbreviation}{Full name of abbreviated variables}
}
\value{
A diagnostic table with/without ROC plot
}
\description{
A diagnostic test output table with/without ROC Curve
}
\examples{
med_marker <- list(
Cholesterol   = list(label = "Serum Cholesterol (Cutoff: 200 mg/dL)", cutoff = 200),
BMI           = list(label = "Body Mass Index (Cutoff: 25 kg/m2)", cutoff = 25),
SBP_Left_Arm  = list(label = "Systolic BP - Left (Cutoff: 140 mmHg)", cutoff = 140)
)
diagnostic_tbl(
 data = mjms_data,
 status_var = "Gold_Standard",
 marker_map = med_marker, # list of all index markers
 show_plot = TRUE,
 plot_marker = "Cholesterol", # show plot for Cholesterol
 table_caption = "Diagnostic Accuracy of Cardiovascular Markers.",
 abbreviation = c("AUC = Area Under Curve", "SBP = Systolic Blood Pressure")
)

}
