% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{lm_tbl}
\alias{lm_tbl}
\title{Linear regression test}
\usage{
lm_tbl(
  data,
  outcome_var,
  predictor_vars,
  var_labels = NULL,
  table_caption = "Factors associated with outcome",
  abbreviation = NULL
)
}
\arguments{
\item{data}{Data frame}

\item{outcome_var}{Outcome variable as a string}

\item{predictor_vars}{Predictor variables as a character vector}

\item{var_labels}{New labels for variables in string}

\item{table_caption}{Caption for the table in string}

\item{abbreviation}{Full name of abbreviated variables}
}
\value{
A linear regression output table
}
\description{
A Linear regression output table for SLR or MLR
}
\examples{
lm_tbl(
  data = mjms_data,
  outcome_var = "Cholesterol",
  predictor_vars = c("Age", "BMI", "Smoker"),
  var_labels = list(Age ~ "Age (Years)", BMI ~ "Body Mass Index"),
  table_caption = "Multiple Linear Regression for Cholesterol"
)

}
