% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{pearson_tbl}
\alias{pearson_tbl}
\title{Pearson's correlation}
\usage{
pearson_tbl(
  data,
  included_var = NULL,
  show_plot = F,
  plot_x = NULL,
  plot_y = NULL,
  abbreviation = NULL,
  table_caption = ""
)
}
\arguments{
\item{data}{Data frame}

\item{included_var}{Variable to be included from the data frame}

\item{show_plot}{TRUE(plot) FALSE(no plot)}

\item{plot_x}{Variable to be x-axis}

\item{plot_y}{Variable to be y-axis}

\item{abbreviation}{Full name of abbreviated variables}

\item{table_caption}{Caption for the table in string}
}
\value{
A Pearson's Correlation output table
}
\description{
A Pearson's correlation output table with/without a correlation plot
}
\examples{
pearson_tbl(
  data = mjms_data,
  included_var = c(Age, BMI, SBP_Left_Arm, SBP_Right_Arm),
  show_plot = TRUE,
  plot_x = "SBP_Left_Arm",
  plot_y = "SBP_Right_Arm",
  abbreviation = "BMI = body mass index, SBP = systolic blood pressure",
  table_caption = "Correlation between variables."
)

}
